feat: adds aragon sponsor and separates into different file
